refactor(audit): split the adhoc shortcut half out of ShortcutManager
shortcutmanager.cpp crossed the 1150-line hard ceiling in this branch, so
the grandfathering clause for existing overruns does not cover it. The
adhoc bindings move to shortcutmanager_adhoc.cpp: transient shortcuts a
caller registers at runtime and owns for as long as its surface is up,
which is a real seam rather than a line count, since that lifetime is
why the block carries its own collision guard and pending-op queue. The
remaining file is 985 lines, under the target rather than merely under
the ceiling. Three test targets link these sources directly and gained
the new TU.

Also: hoist the per-iteration QString off the wheel dispatch loop, note
why the ordinal walk compares tile indices like its two siblings rather
than the active window id, record the live-session check and the
ISO_Left_Tab shape for the Shift'd Tab default, rename a test that
claimed to count drawn tabs when what it pins is the skip, restore the
smoke suite's declaration order to match its definitions, and say why
the work-area shrink there is deliberately not transposed.
